// src/app/models/product-request.ts
export type ProductRequest = {
    id: number
    itemName: string
    quantity: any
    price: number
    dateOfOrder: Date
    customerName: string
    customerEmail: string
    customerPhone: string
    citation: string
    doorNo: any
    street: string
    city: string
    state: string
    zipCode: any
    totalAmount: number
}